@@ -2,22 +2,28 @@
   "name": "triage-desktop",
   "private": true,
   "version": "0.1.0",
+  "type": "module",
+  "main": "./out/main/index.js",
   "scripts": {
-    "dev": "vite",
-    "start": "pnpm run compile && vite & (wait-on http://localhost:5174 && cross-env electron dist-electron/electron/main.js)",
-    "compile": "tsc -p tsconfig.electron.json && tsc -p tsconfig.preload.json",
-    "build": "pnpm lint:check && vite build",
-    "package": "pnpm run build && electron-builder",
-    "lint:check": "eslint src --ext ts,tsx --max-warnings 99",
-    "lint:fix": "eslint src --ext ts,tsx --fix --max-warnings 99",
-    "check-types": "tsc --noEmit",
-    "preview": "vite preview",
-    "typecheck": "tsc --noEmit",
+    "dev": "dotenv -e ../../.env -- electron-vite dev --watch",
+    "start": "dotenv -e ../../.env -- electron-vite preview",
+    "build": "npm run typecheck && electron-vite build",
+    "build:unpack": "npm run build && electron-builder --dir",
+    "build:win": "npm run build && electron-builder --win",
+    "build:mac": "electron-vite build && electron-builder --mac",
+    "build:linux": "electron-vite build && electron-builder --linux",
+    "typecheck": "npm run typecheck:node && npm run typecheck:web",
+    "typecheck:node": "tsc --noEmit -p tsconfig.node.json --composite false",
+    "typecheck:web": "tsc --noEmit -p tsconfig.web.json --composite false",
     "db:generate": "drizzle-kit generate:sqlite",
     "db:migrate": "electron ./scripts/db-migrate-electron.js",
-    "postinstall": "node ./scripts/fix-better-sqlite3.js"
+    "format": "prettier --write .",
+    "lint": "eslint --cache .",
+    "postinstall": "electron-builder install-app-deps"
   },
   "dependencies": {
+    "@electron-toolkit/preload": "^3.0.1",
+    "@electron-toolkit/utils": "^4.0.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
@@ -40,20 +46,27 @@
     "react-split-pane": "^0.1.92"
   },
   "devDependencies": {
-    "@electron/rebuild": "^4.0.1",
+    "@electron-toolkit/eslint-config-prettier": "^3.0.0",
+    "@electron-toolkit/eslint-config-ts": "^3.0.0",
+    "@electron-toolkit/tsconfig": "^1.0.1",
     "@types/better-sqlite3": "^7.6.13",
+    "@types/node": "^22.14.1",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
-    "@vitejs/plugin-react": "^4.2.1",
+    "@vitejs/plugin-react": "^4.3.4",
     "cross-env": "^7.0.3",
-    "electron": "^35.1.0",
-    "electron-builder": "^24.9.1",
-    "electron-rebuild": "^3.2.9",
+    "dotenv-cli": "^8.0.0",
+    "electron": "^35.1.5",
+    "electron-builder": "^25.1.8",
+    "electron-vite": "^3.1.0",
+    "eslint": "^9.24.0",
+    "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
     "prebuild-install": "^7.1.3",
     "ts-node": "^10.9.2",
-    "typescript": "^5.3.3",
-    "vite": "^5.1.0",
+    "typescript": "^5.8.3",
+    "vite": "^6.2.6",
     "wait-on": "^7.2.0"
   }
 }
